# spec/swagger.json
{
  "consumes": [
    "application/json",
    "application/x-www-form-urlencoded"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Welcome to the Ory Kratos HTTP API documentation!",
    "title": "Ory Kratos",
    "version": "latest"
  },
  "basePath": "/",
  "paths": {
...
    "/self-service/login": {
      "post": {
        "description": ":::info\n\nThis endpoint is EXPERIMENTAL and subject to potential breaking changes in the future.\n\n:::\n\nUse this endpoint to complete a login flow. This endpoint\nbehaves differently for API and browser flows.\n\nAPI flows expect `application/json` to be sent in the body and responds with\nHTTP 200 and a application/json body with the session token on success;\nHTTP 302 redirect to a fresh login flow if the original flow expired with the appropriate error messages set;\nHTTP 400 on form validation errors.\n\nBrowser flows expect a Content-Type of `application/x-www-form-urlencoded` or `application/json` to be sent in the body and respond with\na HTTP 302 redirect to the post/after login URL or the `return_to` value if it was set and if the login succeeded;\na HTTP 302 redirect to the login UI URL with the flow ID containing the validation errors otherwise.\n\nBrowser flows with an accept header of `application/json` will not redirect but instead respond with\nHTTP 200 and a application/json body with the signed in identity and a `Set-Cookie` header on success;\nHTTP 302 redirect to a fresh login flow if the original flow expired with the appropriate error messages set;\nHTTP 400 on form validation errors.\n\nMore information can be found at [Ory Kratos User Login and User Registration Documentation](https://www.ory.sh/docs/next/kratos/self-service/flows/user-login-user-registration).",
        "consumes": [
          "application/json",
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json",
          "Header:",
          "Set-Cookie"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "v0alpha2"
        ],
        "summary": "Submit a Login Flow",
        "operationId": "submitSelfServiceLoginFlow",
        "parameters": [
          {
            "type": "string",
            "description": "The Login Flow ID\n\nThe value for this parameter comes from `flow` URL Query parameter sent to your\napplication (e.g. `/login?flow=abcde`).",
            "name": "flow",
            "in": "query",
            "required": true
          },
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/submitSelfServiceLoginFlowBody"
            }
          },
          {
            "type": "string",
            "description": "The Session Token of the Identity performing the settings flow.",
            "name": "X-Session-Token",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "successfulSelfServiceLoginWithoutBrowser",
            "schema": {
              "$ref": "#/definitions/successfulSelfServiceLoginWithoutBrowser"
            }
          },
          "302": {
            "$ref": "#/responses/emptyResponse"
          },
          "400": {
            "description": "selfServiceLoginFlow",
            "schema": {
              "$ref": "#/definitions/selfServiceLoginFlow"
            }
          },
          "500": {
            "description": "jsonError",
            "schema": {
              "$ref": "#/definitions/jsonError"
            }
          }
        }
      }
    },
    "/self-service/login/api": {
      "get": {
        "description": "This endpoint initiates a login flow for API clients that do not use a browser, such as mobile devices, smart TVs, and so on.\n\nIf a valid provided session cookie or session token is provided, a 400 Bad Request error\nwill be returned unless the URL query parameter `?refresh=true` is set.\n\nTo fetch an existing login flow call `/self-service/login/flows?flow=\u003cflow_id\u003e`.\n\nYou MUST NOT use this endpoint in client-side (Single Page Apps, ReactJS, AngularJS) nor server-side (Java Server\nPages, NodeJS, PHP, Golang, ...) browser applications. Using this endpoint in these applications will make\nyou vulnerable to a variety of CSRF attacks, including CSRF login attacks.\n\nThis endpoint MUST ONLY be used in scenarios such as native mobile apps (React Native, Objective C, Swift, Java, ...).\n\nMore information can be found at [Ory Kratos User Login and User Registration Documentation](https://www.ory.sh/docs/next/kratos/self-service/flows/user-login-user-registration).",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "v0alpha2"
        ],
        "summary": "Initialize Login Flow for APIs, Services, Apps, ...",
        "operationId": "initializeSelfServiceLoginFlowWithoutBrowser",
        "parameters": [
          {
            "type": "boolean",
            "description": "Refresh a login session\n\nIf set to true, this will refresh an existing login session by\nasking the user to sign in again. This will reset the\nauthenticated_at time of the session.",
            "name": "refresh",
            "in": "query"
          },
          {
            "type": "string",
            "description": "Request a Specific AuthenticationMethod Assurance Level\n\nUse this parameter to upgrade an existing session's authenticator assurance level (AAL). This\nallows you to ask for multi-factor authentication. When an identity sign in using e.g. username+password,\nthe AAL is 1. If you wish to \"upgrade\" the session's security by asking the user to perform TOTP / WebAuth/ ...\nyou would set this to \"aal2\".",
            "name": "aal",
            "in": "query"
          },
          {
            "type": "string",
            "description": "The Session Token of the Identity performing the settings flow.",
            "name": "X-Session-Token",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "selfServiceLoginFlow",
            "schema": {
              "$ref": "#/definitions/selfServiceLoginFlow"
            }
          },
          "400": {
            "description": "jsonError",
            "schema": {
              "$ref": "#/definitions/jsonError"
            }
          },
          "500": {
            "description": "jsonError",
            "schema": {
              "$ref": "#/definitions/jsonError"
            }
          }
        }
      }
    },
    "/self-service/login/browser": {
      "get": {
        "description": "This endpoint initializes a browser-based user login flow. This endpoint will set the appropriate\ncookies and anti-CSRF measures required for browser-based flows.\n\nIf this endpoint is opened as a link in the browser, it will be redirected to\n`selfservice.flows.login.ui_url` with the flow ID set as the query parameter `?flow=`. If a valid user session\nexists already, the browser will be redirected to `urls.default_redirect_url` unless the query parameter\n`?refresh=true` was set.\n\nIf this endpoint is called via an AJAX request, the response contains the login flow without a redirect.\n\nThis endpoint is NOT INTENDED for clients that do not have a browser (Chrome, Firefox, ...) as cookies are needed.\n\nMore information can be found at [Ory Kratos User Login and User Registration Documentation](https://www.ory.sh/docs/next/kratos/self-service/flows/user-login-user-registration).",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "v0alpha2"
        ],
        "summary": "Initialize Login Flow for Browsers",
        "operationId": "initializeSelfServiceLoginFlowForBrowsers",
        "parameters": [
          {
            "type": "boolean",
            "description": "Refresh a login session\n\nIf set to true, this will refresh an existing login session by\nasking the user to sign in again. This will reset the\nauthenticated_at time of the session.",
            "name": "refresh",
            "in": "query"
          },
          {
            "type": "string",
            "description": "Request a Specific AuthenticationMethod Assurance Level\n\nUse this parameter to upgrade an existing session's authenticator assurance level (AAL). This\nallows you to ask for multi-factor authentication. When an identity sign in using e.g. username+password,\nthe AAL is 1. If you wish to \"upgrade\" the session's security by asking the user to perform TOTP / WebAuth/ ...\nyou would set this to \"aal2\".",
            "name": "aal",
            "in": "query"
          },
          {
            "type": "string",
            "description": "The Session Token of the Identity performing the settings flow.",
            "name": "X-Session-Token",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "selfServiceLoginFlow",
            "schema": {
              "$ref": "#/definitions/selfServiceLoginFlow"
            }
          },
          "302": {
            "$ref": "#/responses/emptyResponse"
          },
          "500": {
            "description": "jsonError",
            "schema": {
              "$ref": "#/definitions/jsonError"
            }
          }
        }
      }
    },
    "/self-service/login/flows": {
      "get": {
        "description": "This endpoint returns a login flow's context with, for example, error details and other information.\n\nBrowser flows expect the anti-CSRF cookie to be included in the request's HTTP Cookie Header.\nFor AJAX requests you must ensure that cookies are included in the request or requests will fail.\n\nIf you use the browser-flow for server-side apps, the services need to run on a common top-level-domain\nand you need to forward the incoming HTTP Cookie header to this endpoint:\n\n```js\npseudo-code example\nrouter.get('/login', async function (req, res) {\nconst flow = await client.getSelfServiceLoginFlow(req.header('cookie'), req.query['flow'])\n\nres.render('login', flow)\n})\n```\n\nMore information can be found at [Ory Kratos User Login and User Registration Documentation](https://www.ory.sh/docs/next/kratos/self-service/flows/user-login-user-registration).",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "v0alpha2"
        ],
        "summary": "Get Login Flow",
        "operationId": "getSelfServiceLoginFlow",
        "parameters": [
          {
            "type": "string",
            "description": "The Login Flow ID\n\nThe value for this parameter comes from `flow` URL Query parameter sent to your\napplication (e.g. `/login?flow=abcde`).",
            "name": "id",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "description": "HTTP Cookies\n\nWhen using the SDK on the server side you must include the HTTP Cookie Header\noriginally sent to your HTTP handler here.",
            "name": "cookie",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "selfServiceLoginFlow",
            "schema": {
              "$ref": "#/definitions/selfServiceLoginFlow"
            }
          },
          "403": {
            "description": "jsonError",
            "schema": {
              "$ref": "#/definitions/jsonError"
            }
          },
          "404": {
            "description": "jsonError",
            "schema": {
              "$ref": "#/definitions/jsonError"
            }
          },
          "410": {
            "description": "jsonError",
            "schema": {
              "$ref": "#/definitions/jsonError"
            }
          },
          "500": {
            "description": "jsonError",
            "schema": {
              "$ref": "#/definitions/jsonError"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "UUID": {"type": "string", "format": "uuid4"},
...
    "identityCredentialsType": {
      "description": "and so on.",
      "type": "string",
      "title": "CredentialsType  represents several different credential types, like password credentials, passwordless credentials,"
    },
...
    "jsonError": {
      "description": "The standard Ory JSON API error format.",
      "type": "object",
      "title": "JSON API Error Response",
      "required": [
        "error"
      ],
      "properties": {
        "error": {
          "$ref": "#/definitions/genericError"
        }
      }
    },
...
    "selfServiceFlowType": {
      "description": "The flow type can either be `api` or `browser`.",
      "type": "string",
      "title": "Type is the flow type."
    },
    "selfServiceLoginFlow": {
      "description": "This object represents a login flow. A login flow is initiated at the \"Initiate Login API / Browser Flow\"\nendpoint by a client.\n\nOnce a login flow is completed successfully, a session cookie or session token will be issued.",
      "type": "object",
      "title": "Login Flow",
      "required": [
        "id",
        "type",
        "expires_at",
        "issued_at",
        "request_url",
        "ui"
      ],
      "properties": {
        "active": {
          "$ref": "#/definitions/identityCredentialsType"
        },
        "created_at": {
          "description": "CreatedAt is a helper struct field for gobuffalo.pop.",
          "type": "string",
          "format": "date-time"
        },
        "expires_at": {
          "description": "ExpiresAt is the time (UTC) when the flow expires. If the user still wishes to log in,\na new flow has to be initiated.",
          "type": "string",
          "format": "date-time"
        },
        "forced": {
          "description": "Refresh stores whether this login flow should enforce re-authentication.",
          "type": "boolean"
        },
        "id": {
          "$ref": "#/definitions/UUID"
        },
        "issued_at": {
          "description": "IssuedAt is the time (UTC) when the flow started.",
          "type": "string",
          "format": "date-time"
        },
        "request_url": {
          "description": "RequestURL is the initial URL that was requested from Ory Kratos. It can be used\nto forward information contained in the URL's path or query for example.",
          "type": "string"
        },
        "requested_aal": {
          "$ref": "#/definitions/authenticatorAssuranceLevel"
        },
        "type": {
          "$ref": "#/definitions/selfServiceFlowType"
        },
        "ui": {
          "$ref": "#/definitions/uiContainer"
        },
        "updated_at": {
          "description": "UpdatedAt is a helper struct field for gobuffalo.pop.",
          "type": "string",
          "format": "date-time"
        }
      }
    },
...
    "session": {
      "description": "A Session",
      "type": "object",
      "required": [
        "id",
        "identity"
      ],
      "properties": {
        "active": {
          "description": "Active state. If false the session is no longer active.",
          "type": "boolean"
        },
        "authenticated_at": {
          "description": "The Session Authentication Timestamp\n\nWhen this session was authenticated at. If multi-factor authentication was used this\nis the time when the last factor was authenticated (e.g. the TOTP code challenge was completed).",
          "type": "string",
          "format": "date-time"
        },
        "authentication_methods": {
          "$ref": "#/definitions/sessionAuthenticationMethods"
        },
        "authenticator_assurance_level": {
          "$ref": "#/definitions/authenticatorAssuranceLevel"
        },
        "expires_at": {
          "description": "The Session Expiry\n\nWhen this session expires at.",
          "type": "string",
          "format": "date-time"
        },
        "id": {
          "$ref": "#/definitions/UUID"
        },
        "identity": {
          "$ref": "#/definitions/identity"
        },
        "issued_at": {
          "description": "The Session Issuance Timestamp\n\nWhen this session was issued at. Usually equal or close to `authenticated_at`.",
          "type": "string",
          "format": "date-time"
        }
      }
    },
...
    "sessionAuthenticationMethods": {
      "description": "A list of authenticators which were used to authenticate the session.",
      "type": "array",
      "title": "List of (Used) AuthenticationMethods",
      "items": {
        "$ref": "#/definitions/sessionAuthenticationMethod"
      }
    },
...
    "submitSelfServiceLoginFlowBody": {
      "type": "object"
    },
...
    "successfulSelfServiceLoginWithoutBrowser": {
      "description": "The Response for Login Flows via API",
      "type": "object",
      "required": [
        "session"
      ],
      "properties": {
        "session": {
          "$ref": "#/definitions/session"
        },
        "session_token": {
          "description": "The Session Token\n\nA session token is equivalent to a session cookie, but it can be sent in the HTTP Authorization\nHeader:\n\nAuthorization: bearer ${session-token}\n\nThe session token is only issued for API flows, not for Browser flows!",
          "type": "string"
        }
      }
    },
...
  },
  "responses": {
    "emptyResponse": {
      "description": "Empty responses are sent when, for example, resources are deleted. The HTTP status code for empty responses is typically 201."
    }
  },
  "securityDefinitions": {
    "oryAccessToken": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
...
}